feat(workbench): progressive disclosure for AI and organization guards for Team Curation Queue
This implements P0 priorities from the May 21st App Audit:
- Treat 'heuristic' and 'local' AI providers as disabled to hide unconfigured/unsupported semantic/hybrid search and RAG options
- Implement activeOrgId check on TeamReviewPage to restrict access to Team Curation Queue when the user is not in an organization
- Refactor monolithic WorkbenchPage into modular, cohesive tool components under packages/features
